AI_Tech_Consultant_Agent/pyproject.toml
[tool.poetry]
name = "ai-tech-consultant-agent"
version = "0.1.0"
description = "AI Tech Consultant Agent using Google ADK and Mem0"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-generativeai = "^0.8.5"
google-adk = "^1.6.1"
mem0ai = "^0.1.114"
litellm = "^1.74.2"
langfuse = "^3.1.3"
python-dotenv = "^1.1.1"
opentelemetry-sdk = "^1.35.0"
opentelemetry-exporter-otlp = "^1.35.0"
chromadb = "^1.0.15"
adk = "^0.0.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
multi_line_output = 3 